requirements.dev.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
flake8==7.0.0
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==24.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
uritemplate==4.1.1
